+++ src/client/components/teScriber/teScriber.js
@@ -37,7 +37,7 @@ Template.teScriber.onCreated( function(){
         // the parms
         content: null,
         name: '',
-        currentMode: new ReactiveVar( TE_MODE_STANDARD ),
+        currentMode: new ReactiveVar( Editor.C.Mode.STANDARD ),
         displayName: true,
         withNamePanel: true,
         withHTMLBtn: true,
@@ -73,16 +73,16 @@ Template.teScriber.onCreated( function(){
                 _switchState = Editor.switch.state.get();
                 if( !_switchState ){
                     switch( mode ){
-                        case TE_MODE_PREVIEW:
-                        case TE_MODE_EDITION:
-                            mode =  TE_MODE_STANDARD;
-                            self.TE.mode_asked = TE_MODE_PREVIEW;
+                        case Editor.C.Mode.PREVIEW:
+                        case Editor.C.Mode.EDITION:
+                            mode =  Editor.C.Mode.STANDARD;
+                            self.TE.mode_asked = Editor.C.Mode.PREVIEW;
                             break;
                     }
                 }
 
             }
-            if( Editor._conf.verbosity & TE_VERBOSE_MODE ){
+            if( Editor._conf.verbosity & Editor.C.Verbose.MODE ){
                 console.debug( 'pwix:editor teScriber checkSwitch() asked='+_input, 'switchExists='+ _hasSwitch, 'switchState='+_switchState, 'returning', mode );
             }
             return mode;
@@ -157,7 +157,7 @@ Template.teScriber.onCreated( function(){
         //  this code run the creation; the initialization itself is made on tbwinit message
         editorCreate(){
             if( self.view.isRendered && !self.TE.editorInitialized.get()){
-                if( Editor._conf.verbosity & TE_VERBOSE_TRUMBOWYG ){
+                if( Editor._conf.verbosity & Editor.C.Verbose.TRUMBOWYG ){
                     console.debug( 'pwi:editor teScriber editorCreate() instanciating...' );
                 }
                 self.TE.editorDiv.trumbowyg({
@@ -174,7 +174,7 @@ Template.teScriber.onCreated( function(){
         // delete the Trumbowyg edtor (quitting the EDITION mode)
         editorDelete(){
             if( self.TE.editorInitialized.get()){
-                if( Editor._conf.verbosity & TE_VERBOSE_TRUMBOWYG ){
+                if( Editor._conf.verbosity & Editor.C.Verbose.TRUMBOWYG ){
                     console.debug( 'pwix:editor teScriber editorDelete() destroying instance' );
                 }
                 const res = self.TE.editorDiv.trumbowyg( 'destroy' );
@@ -204,14 +204,14 @@ Template.teScriber.onCreated( function(){
         // get/set the edit mode
         mode( mode=null ){
             if( mode ){
-                if( Editor._conf.verbosity & TE_VERBOSE_MODE ){
+                if( Editor._conf.verbosity & Editor.C.Verbose.MODE ){
                     console.debug( 'pwix:editor teScriber mode() asked='+mode );
                 }
                 self.TE.mode_asked = mode;
                 if( Editor.Modes.includes( mode )){
                     mode = self.TE.checkSwitch( mode );
                     const prev = self.TE.currentMode.get();
-                    if( Editor._conf.verbosity & TE_VERBOSE_MODE ){
+                    if( Editor._conf.verbosity & Editor.C.Verbose.MODE ){
                         console.debug( 'pwix:editor teScriber mode() previous was', prev );
                     }
                     if( prev !== mode ){
@@ -219,7 +219,7 @@ Template.teScriber.onCreated( function(){
                             self.$( '.teScriber' ).trigger( 'te-mode-changed', { prev: prev, new: mode });
                         }
                         self.TE.currentMode.set( mode );
-                    } else if( Editor._conf.verbosity & TE_VERBOSE_MODE ){
+                    } else if( Editor._conf.verbosity & Editor.C.Verbose.MODE ){
                         console.debug( 'pwix:editor teScriber mode() mode is unchanged' );
                     }
                 } else {
@@ -264,7 +264,7 @@ Template.teScriber.onCreated( function(){
     self.autorun(() => {
         const mode = Template.currentData().mode;
         if( mode !== self.TE.mode_parm ){
-            if( Editor._conf.verbosity & TE_VERBOSE_MODE ){
+            if( Editor._conf.verbosity & Editor.C.Verbose.MODE ){
                 console.debug( 'pwix:editor teScriber currentData() asking mode='+mode );
             }
             self.TE.mode_parm = mode;
@@ -295,7 +295,7 @@ Template.teScriber.onCreated( function(){
     // upload url for the images
     self.autorun(() => {
         self.TE.uploadUrl = Editor._conf.uploadUrl;
-        if( Editor._conf.verbosity & TE_VERBOSE_UPLOAD ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.UPLOAD ){
             console.debug( 'pwix:editor teScriber uploadUrl', self.TE.uploadUrl );
         }
     });
@@ -306,7 +306,7 @@ Template.teScriber.onCreated( function(){
         const _hasSwitch = Editor.switch.used.get();
         if( _hasSwitch ){
             const _state = Editor.switch.state.get();
-            if( Editor._conf.verbosity & TE_VERBOSE_MODE ){
+            if( Editor._conf.verbosity & Editor.C.Verbose.MODE ){
                 console.debug( 'pwix:editor teScriber auto follow teSwitch for', self.TE.mode_asked );
             }
             self.TE.mode( self.TE.mode_asked );
@@ -314,7 +314,7 @@ Template.teScriber.onCreated( function(){
     });
 
     // be verbose
-    if( Editor._conf.verbosity & TE_VERBOSE_COMPONENTS ){
+    if( Editor._conf.verbosity & Editor.C.Verbose.COMPONENTS ){
         console.debug( 'pwix:editor teScriber onCreated()', self.TE.name );
     }
 });
@@ -323,7 +323,7 @@ Template.teScriber.onRendered( function(){
     const self = this;
 
     // be verbose
-    if( Editor._conf.verbosity & TE_VERBOSE_COMPONENTS ){
+    if( Editor._conf.verbosity & Editor.C.Verbose.COMPONENTS ){
         console.debug( 'pwix:editor teScriber onRendered()' );
     }
 
@@ -340,24 +340,24 @@ Template.teScriber.onRendered( function(){
     // honor the ask mode
     self.autorun(() => {
         const mode = self.TE.mode();
-        if( Editor._conf.verbosity & TE_VERBOSE_MODE ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.MODE ){
             console.debug( 'pwix:editor teScriber autorun() honoring', mode );
         }
         self.$( '.te-edit-container' ).removeClass().addClass( 'te-edit-container '+mode );
         switch( mode ){
-            case TE_MODE_HIDDEN:
+            case Editor.C.Mode.HIDDEN:
                 self.TE.editorDelete();
                 break;
-            case TE_MODE_STANDARD:
+            case Editor.C.Mode.STANDARD:
                 self.TE.editorDelete();
                 self.TE.contentHtml();
                 break;
-            case TE_MODE_PREVIEW:
+            case Editor.C.Mode.PREVIEW:
                 self.TE.editorDelete();
                 self.TE.contentHtml();
                 self.TE.toggleOn( '.te-view-btn' );
                 break;
-            case TE_MODE_EDITION:
+            case Editor.C.Mode.EDITION:
                 self.TE.editorCreate();
                 self.TE.toggleOn( '.te-edit-btn' );
                 break;
@@ -381,13 +381,13 @@ Template.teScriber.helpers({
 Template.teScriber.events({
     // change the edition mode
     'click .te-edit-btn'( event, instance ){
-        instance.TE.mode( TE_MODE_EDITION );
+        instance.TE.mode( Editor.C.Mode.EDITION );
         // doesn't propagate the event
         return false;
     },
 
     'click .te-view-btn'( event, instance ){
-        instance.TE.mode( TE_MODE_PREVIEW );
+        instance.TE.mode( Editor.C.Mode.PREVIEW );
         // doesn't propagate the event
         return false;
     },
@@ -397,7 +397,7 @@ Template.teScriber.events({
         const editorInitialized = instance.TE.editorInitialized.get();
         const tbwchange_update = instance.TE.tbwchange_update;
         const tbwchange_sendmsg = instance.TE.tbwchange_sendmsg;
-        if( Editor._conf.verbosity & TE_VERBOSE_TRUMBOWYG ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.TRUMBOWYG ){
             console.debug( 'pwix:editor teScriber tbwchange editorInitialized='+editorInitialized, 'tbwchange_update='+tbwchange_update, 'tbwchange_sendmsg='+tbwchange_sendmsg );
         }
         if( editorInitialized ){
@@ -410,17 +410,17 @@ Template.teScriber.events({
                         o.name = Template.currentData().name;
                     }
                     instance.$( '.teScriber' ).trigger( 'te-content-changed', o );
-                } else if( Editor._conf.verbosity & TE_VERBOSE_TRUMBOWYG ){
+                } else if( Editor._conf.verbosity & Editor.C.Verbose.TRUMBOWYG ){
                     console.debug( 'pwix:editor teScriber tbwchange tbwchange_sendmsg is false' );
                 }
                 // update the input/output ReactiveVar if allowed to
                 if( tbwchange_update && instance.TE.content ){
                     instance.TE.content.set( html );
-                } else if( Editor._conf.verbosity & TE_VERBOSE_TRUMBOWYG ){
+                } else if( Editor._conf.verbosity & Editor.C.Verbose.TRUMBOWYG ){
                     console.debug( 'pwix:editor teScriber tbwchange tbwchange_update is false' );
                 }
                 instance.TE.tbwchange_last = html;
-            } else if( Editor._conf.verbosity & TE_VERBOSE_TRUMBOWYG ){
+            } else if( Editor._conf.verbosity & Editor.C.Verbose.TRUMBOWYG ){
                 console.debug( 'pwix:editor teScriber tbwchange html is unchanged' );
             }
         }
@@ -428,14 +428,14 @@ Template.teScriber.events({
 
     // the editor is destroyed
     'tbwclose .te-edit-content'( event, instance ){
-        if( Editor._conf.verbosity & TE_VERBOSE_TRUMBOWYG ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.TRUMBOWYG ){
             console.debug( 'pwix:editor teScriber tbwclose' );
         }
     },
 
     // the editor is initialized
     'tbwinit .te-edit-content'( event, instance ){
-        if( Editor._conf.verbosity & TE_VERBOSE_TRUMBOWYG ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.TRUMBOWYG ){
             console.debug( 'pwix:editor teScriber tbwinit' );
         }
         instance.TE.editorInitialized.set( true );
@@ -452,14 +452,14 @@ Template.teScriber.events({
 
     // html content has changed
     'te-content-changed .teScriber'( event, instance, data ){
-        if( Editor._conf.verbosity & TE_VERBOSE_TEMSG ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.TEMSG ){
             console.debug( 'pwix:editor teScriber te-content-changed', data );
         }
     },
 
     // re-set the html content of the editing area
     'te-content-reset .teScriber'( event, instance ){
-        if( Editor._conf.verbosity & TE_VERBOSE_TEMSG ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.TEMSG ){
             console.debug( 'pwix:editor teScriber te-content-reset' );
         }
         instance.TE.contentReset();
@@ -469,14 +469,14 @@ Template.teScriber.events({
     // debug the te-initialized message
     //  let event bubble up
     'te-initialized .teScriber'( event, instance ){
-        if( Editor._conf.verbosity & TE_VERBOSE_TEMSG ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.TEMSG ){
             console.debug( 'pwix:editor teScriber te-initialized' );
         }
     },
 
     // request to change the edition mode
     'te-mode-set .teScriber'( event, instance, data ){
-        if( Editor._conf.verbosity & TE_VERBOSE_TEMSG ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.TEMSG ){
             console.debug( 'pwix:editor teScriber te-mode-set', data );
         }
         instance.TE.mode( data.mode );
@@ -486,7 +486,7 @@ Template.teScriber.events({
     // the edition mode has changed
     //  let the event bubble up
     'te-mode-changed .teScriber'( event, instance, data ){
-        if( Editor._conf.verbosity & TE_VERBOSE_TEMSG ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.TEMSG ){
             console.debug( 'pwix:editor teScriber te-mode-changed', data );
         }
     }
@@ -496,7 +496,7 @@ Template.teScriber.onDestroyed( function(){
     this.TE.editorDelete();
 
     // be verbose
-    if( Editor._conf.verbosity & TE_VERBOSE_COMPONENTS ){
+    if( Editor._conf.verbosity & Editor.C.Verbose.COMPONENTS ){
         console.debug( 'pwix:editor teScriber onDestroyed()' );
     }
 });