+++ src/collections/contents/server/methods.js
@@ -3,11 +3,11 @@ import { Contents } from '../contents.js';
 Meteor.methods({
     // empty the collection
     'te_contents.empty'(){
-        if( Editor._conf.verbosity & TE_VERBOSE_COLLECTIONS ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.empty() call' );
         }
         const res = Editor.collections.Contents.server.remove({});
-        if( Editor._conf.verbosity & TE_VERBOSE_COLLECTIONS ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.empty() returns with', res );
         }
         return res;
@@ -15,11 +15,11 @@ Meteor.methods({
 
     // get some content from the database
     'te_contents.byName'( name ){
-        if( Editor._conf.verbosity & TE_VERBOSE_COLLECTIONS ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.byName() call with name=', name );
         }
         const res = Editor.collections.Contents.server.findOne({ name: name });
-        if( Editor._conf.verbosity & TE_VERBOSE_COLLECTIONS ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.byName() returns with', res );
         }
         return res;
@@ -27,11 +27,11 @@ Meteor.methods({
 
     // import an element (so do not modify any data) 
     'te_contents.import'( elt ){
-        if( Editor._conf.verbosity & TE_VERBOSE_COLLECTIONS ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.import() call with elt=', elt );
         }
         const res = Editor.collections.Contents.server.insert( elt );
-        if( Editor._conf.verbosity & TE_VERBOSE_COLLECTIONS ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.import() returns with', res );
         }
         return res;
@@ -40,7 +40,7 @@ Meteor.methods({
     // set some content into the database
     // name is mandatory
     'te_contents.set'( name, content ){
-        if( Editor._conf.verbosity & TE_VERBOSE_COLLECTIONS ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.set() call with name=', name, 'content=', content );
         }
         if( !name ){
@@ -66,7 +66,7 @@ Meteor.methods({
         if( res.numberAffected > 0 ){
             res.written = Editor.collections.Contents.server.findOne({ name: name });
         }
-        if( Editor._conf.verbosity & TE_VERBOSE_COLLECTIONS ){
+        if( Editor._conf.verbosity & Editor.C.Verbose.COLLECTIONS ){
             console.debug( 'pwix:editor te_contents.set() returns with', res );
         }
         return res;