+++ src/common/js/configure.js
@@ -10,7 +10,7 @@ Editor._defaults = {
     },
     storeSwitchState: false,
     uploadUrl: '',
-    verbosity: TE_VERBOSE_NONE
+    verbosity: Editor.C.Verbose.NONE
 };
 
 /**
@@ -24,7 +24,7 @@ Editor.configure = function( o ){
     if( o && _.isObject( o )){
         _.merge( Editor._conf, Editor._defaults, o );
     }
-    if( Editor._conf.verbosity & TE_VERBOSE_CONFIGURE ){
+    if( Editor._conf.verbosity & Editor.C.Verbose.CONFIGURE ){
         console.debug( 'pwix:editor configure() with', o, 'building', Editor._conf );
     }
     // also acts as a getter