+++ src/common/js/constants.js
@@ -2,37 +2,38 @@
  * pwix:editor/src/common/js/constants.js
  */
 
-// internal constants
-//
+Editor.C = {
+    // display mode
+    Mode: {
+        HIDDEN:   'MODE_HIDDEN',
+        STANDARD: 'MODE_STANDARD',
+        PREVIEW:  'MODE_PREVIEW',
+        EDITION:  'MODE_EDITION'
+    },
+    // verbosity levels
+    Verbose: {
+        NONE:        0,
+        MODE:        0x01 <<  0,
+        TEMSG:       0x01 <<  1,
+        TRUMBOWYG:   0x01 <<  2,
+        UPLOAD:      0x01 <<  3,
+        CONFIGURE:   0x01 <<  4,
+        COLLECTIONS: 0x01 <<  5,
+        COMPONENTS:  0x01 <<  6,
+        SWITCH:      0x01 <<  7
+    }
+};
+
+Editor.Modes = [];
+Object.keys( Editor.C.Mode ).every(( k ) => {
+    Editor.Modes.push( Editor.C.Mode[k] );
+    return true;
+});
+
+// internal (non exported) constants
 
 // the cookie name for the switch state, defined here to be available anywhere
-COOKIE_SWITCH_STATE = 'pwix:Editor/switch_state';
-
-// exported constants
-//
-
-TE_MODE_HIDDEN = 'TE_MODE_HIDDEN';
-TE_MODE_STANDARD = 'TE_MODE_STANDARD';
-TE_MODE_PREVIEW = 'TE_MODE_PREVIEW';
-TE_MODE_EDITION = 'TE_MODE_EDITION';
-
-Editor.Modes = [
-    TE_MODE_HIDDEN,
-    TE_MODE_STANDARD,
-    TE_MODE_PREVIEW,
-    TE_MODE_EDITION
-];
-
-TE_VERBOSE_NONE        = 0;
-TE_VERBOSE_MODE        = 0x01 <<  0;
-TE_VERBOSE_TEMSG       = 0x01 <<  1;
-TE_VERBOSE_TRUMBOWYG   = 0x01 <<  2;
-TE_VERBOSE_UPLOAD      = 0x01 <<  3;
-TE_VERBOSE_CONFIGURE   = 0x01 <<  4;
-TE_VERBOSE_COLLECTIONS = 0x01 <<  5;
-TE_VERBOSE_COMPONENTS  = 0x01 <<  6;
-TE_VERBOSE_SWITCH      = 0x01 <<  7;
-
-// non exported
+COOKIE_RESPONSIBLE = 'pwix:Editor';
+COOKIE_SWITCH_STATE = 'switch_state';
 
 I18N = 'pwix:editor:i18n';